[Cleanup] Extract shared ColumnParallelGELU into diffusion/layers/activations
Replace 6 duplicate ColumnParallelGELU/ColumnParallelApproxGELU class
definitions with a single shared implementation. The shared class accepts
approximate, quant_config, and prefix as parameters to cover all variants.

Also fixes qwen_image prefix handling to match the standard convention
(f"{prefix}.net.0.proj" / f"{prefix}.net.2") used by flux.
